tests/wpt/web-platform-tests/webmessaging/without-ports/003.html

<!doctype html>
<title>resolving 'example.org'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws('SYNTAX_ERR', function() {
    postMessage('', 'example.org');
  }, 'targetOrigin is not an absolute url');
});
</script>
